+++ src/components/controlled/transactions-table/components/transaction-shards/transaction-shards.tsx
@@ -4,6 +4,10 @@ import { DataTestIdsEnum } from 'constants/dataTestIds.enum';
 
 import type { TransactionRowType } from '../../transactions-table.type';
 
+const transactionShardsClasses: Record<string, string> = {
+  explorerLink: 'drt:text-blue-link!',
+};
+
 @Component({
   tag: 'drt-transaction-shards',
   styleUrl: 'transaction-shards.scss',
@@ -17,7 +21,7 @@ export class TransactionShards {
       <div class={classNames(this.class, 'transaction-shards')}>
         <drt-explorer-link
           link={this.transaction.sender.shardLink}
-          class="transactions-table-body-cell-link"
+          class={transactionShardsClasses.explorerLink}
           data-testid={DataTestIdsEnum.shardFromLink}
         >
           <span data-testid={DataTestIdsEnum.senderShard}>{this.transaction.sender.shard}</span>
@@ -26,9 +30,9 @@ export class TransactionShards {
         <span class="transaction-shards-arrow">&#10132;</span>
 
         <drt-explorer-link
-          class="transactions-table-body-cell-link"
           link={this.transaction.receiver.shardLink}
           data-testid={DataTestIdsEnum.shardToLink}
+          class={transactionShardsClasses.explorerLink}
         >
           <span data-testid={DataTestIdsEnum.receiverShard}>{this.transaction.receiver.shard}</span>
         </drt-explorer-link>