+++ src/components/functional/unlock-panel/helpers/getProviderButtonIcon.tsx
@@ -1,11 +1,12 @@
 import { h } from '@stencil/core';
 import type { IProviderBase } from 'types/provider.types';
 import { ProviderTypeEnum } from 'types/provider.types';
+import { isFirefox } from 'utils/isFirefox';
 
 export const getProviderButtonIcon = (providerType: IProviderBase['type']): HTMLElement => {
   switch (providerType) {
     case ProviderTypeEnum.extension:
-      return <drt-extension-provider-icon />;
+      return isFirefox() ? <drt-wallet-provider-icon /> : <drt-extension-provider-icon />;
     case ProviderTypeEnum.metamask:
       return <drt-metamask-provider-icon />;
     case ProviderTypeEnum.passkey: