+++ src/utils/isFirefox.ts
@@ -1,0 +1,3 @@
+import { safeWindow } from 'constants/window.constants';
+
+export const isFirefox = () => safeWindow?.navigator?.userAgent.match('Firefox');